fix: various: moved GameBoard{} with RandInt{} and Tile{} to their own .hpp and .cpp
As refactoring was cluttering up game.hpp, this move was needed to help
encapsulation (SRP) of the class functions.
